Is it possible
There are three cards with letters a, b, c placed in a row in some order. You can do the following operation at most once:
Pick two cards, and swap them.
Is it possible that the row becomes abc after the operation? Output "YES" if it is possible, and "NO" otherwise.